+++ config/users.php
@@ -1,12 +1,12 @@
 <?php
 
 /**
- * Copyright 2010 - 2015, Cake Development Corporation (http://cakedc.com)
+ * Copyright 2010 - 2017, Cake Development Corporation (https://www.cakedc.com)
  *
  * Licensed under The MIT License
  * Redistributions of files must retain the above copyright notice.
  *
- * @copyright Copyright 2010 - 2015, Cake Development Corporation (http://cakedc.com)
+ * @copyright Copyright 2010 - 2017, Cake Development Corporation (https://www.cakedc.com)
  * @license MIT License (http://www.opensource.org/licenses/mit-license.php)
  */
 use Cake\Core\Configure;
@@ -14,50 +14,54 @@ use Cake\Routing\Router;
 
 $config = [
     'Users' => [
-        //Table used to manage users
+        // Table used to manage users
         'table' => 'CakeDC/Users.Users',
-        //configure Auth component
+        // Controller used to manage users plugin features & actions
+        'controller' => 'CakeDC/Users.Users',
+        // configure Auth component
         'auth' => true,
-        //Password Hasher
+        // Password Hasher
         'passwordHasher' => '\Cake\Auth\DefaultPasswordHasher',
-        //token expiration, 1 hour
+        // token expiration, 1 hour
         'Token' => ['expiration' => 3600],
         'Email' => [
-            //determines if the user should include email
+            // determines if the user should include email
             'required' => true,
-            //determines if registration workflow includes email validation
+            // determines if registration workflow includes email validation
             'validate' => true,
         ],
         'Registration' => [
-            //determines if the register is enabled
+            // determines if the register is enabled
             'active' => true,
-            //determines if the reCaptcha is enabled for registration
+            // determines if the reCaptcha is enabled for registration
             'reCaptcha' => true,
-            //allow a logged in user to access the registration form
+            // allow a logged in user to access the registration form
             'allowLoggedIn' => false,
             //ensure user is active (confirmed email) to reset his password
-            'ensureActive' => false
+            'ensureActive' => false,
+            // default role name used in registration
+            'defaultRole' => 'user',
         ],
         'reCaptcha' => [
-            //reCaptcha key goes here
+            // reCaptcha key goes here
             'key' => null,
-            //reCaptcha secret
+            // reCaptcha secret
             'secret' => null,
-            //use reCaptcha in registration
+            // use reCaptcha in registration
             'registration' => false,
-            //use reCaptcha in login, valid values are false, true
+            // use reCaptcha in login, valid values are false, true
             'login' => false,
         ],
         'Tos' => [
-            //determines if the user should include tos accepted
+            // determines if the user should include tos accepted
             'required' => true,
         ],
         'Social' => [
-            //enable social login
+            // enable social login
             'login' => false,
         ],
         'GoogleAuthenticator' => [
-            //enable Google Authenticator
+            // enable Google Authenticator
             'login' => false,
             'issuer' => null,
             // The number of digits the resulting codes will be
@@ -72,34 +76,34 @@ $config = [
             'rngprovider' => null
         ],
         'Profile' => [
-            //Allow view other users profiles
+            // Allow view other users profiles
             'viewOthers' => true,
             'route' => ['plugin' => 'CakeDC/Users', 'controller' => 'Users', 'action' => 'profile'],
         ],
         'Key' => [
             'Session' => [
-                //session key to store the social auth data
+                // session key to store the social auth data
                 'social' => 'Users.social',
-                //userId key used in reset password workflow
+                // userId key used in reset password workflow
                 'resetPasswordUserId' => 'Users.resetPasswordUserId',
             ],
-            //form key to store the social auth data
+            // form key to store the social auth data
             'Form' => [
                 'social' => 'social'
             ],
             'Data' => [
-                //data key to store the users email
+                // data key to store the users email
                 'email' => 'email',
-                //data key to store email coming from social networks
+                // data key to store email coming from social networks
                 'socialEmail' => 'info.email',
-                //data key to check if the remember me option is enabled
+                // data key to check if the remember me option is enabled
                 'rememberMe' => 'remember_me',
             ],
         ],
-        //Avatar placeholder
+        // Avatar placeholder
         'Avatar' => ['placeholder' => 'CakeDC/Users.avatar_placeholder.png'],
         'RememberMe' => [
-            //configure Remember Me component
+            // configure Remember Me component
             'active' => true,
             'checked' => true,
             'Cookie' => [
@@ -119,7 +123,7 @@ $config = [
             'prefix' => false,
         ],
     ],
-    //default configuration used to auto-load the Auth Component, override to change the way Auth works
+    // default configuration used to auto-load the Auth Component, override to change the way Auth works
     'Auth' => [
         'loginAction' => [
             'plugin' => 'CakeDC/Users',
@@ -131,13 +135,13 @@ $config = [
             'all' => [
                 'finder' => 'auth',
             ],
-            'CakeDC/Users.ApiKey',
-            'CakeDC/Users.RememberMe',
+            'CakeDC/Auth.ApiKey',
+            'CakeDC/Auth.RememberMe',
             'Form',
         ],
         'authorize' => [
-            'CakeDC/Users.Superuser',
-            'CakeDC/Users.SimpleRbac',
+            'CakeDC/Auth.Superuser',
+            'CakeDC/Auth.SimpleRbac',
         ],
     ],
     'OAuth' => [