+++ src/Controller/Component/UsersAuthComponent.php
@@ -1,11 +1,11 @@
 <?php
 /**
- * Copyright 2010 - 2015, Cake Development Corporation (http://cakedc.com)
+ * Copyright 2010 - 2017, Cake Development Corporation (https://www.cakedc.com)
  *
  * Licensed under The MIT License
  * Redistributions of files must retain the above copyright notice.
  *
- * @copyright Copyright 2010 - 2015, Cake Development Corporation (http://cakedc.com)
+ * @copyright Copyright 2010 - 2017, Cake Development Corporation (https://www.cakedc.com)
  * @license MIT License (http://www.opensource.org/licenses/mit-license.php)
  */
 
@@ -16,7 +16,7 @@ use Cake\Controller\Component;
 use Cake\Core\Configure;
 use Cake\Event\Event;
 use Cake\Event\EventManager;
-use Cake\Network\Request;
+use Cake\Http\ServerRequest;
 use Cake\Routing\Exception\MissingRouteException;
 use Cake\Routing\Router;
 use Cake\Utility\Hash;
@@ -67,7 +67,7 @@ class UsersAuthComponent extends Component
      */
     protected function _loadGoogleAuthenticator()
     {
-        $this->_registry->getController()->loadComponent('CakeDC/Users.GoogleAuthenticator');
+        $this->getController()->loadComponent('CakeDC/Users.GoogleAuthenticator');
     }
 
     /**
@@ -77,7 +77,7 @@ class UsersAuthComponent extends Component
      */
     protected function _loadSocialLogin()
     {
-        $this->_registry->getController()->Auth->config('authenticate', [
+        $this->getController()->Auth->setConfig('authenticate', [
             'CakeDC/Users.Social'
         ], true);
     }
@@ -89,7 +89,7 @@ class UsersAuthComponent extends Component
      */
     protected function _loadRememberMe()
     {
-        $this->_registry->getController()->loadComponent('CakeDC/Users.RememberMe');
+        $this->getController()->loadComponent('CakeDC/Users.RememberMe');
     }
 
     /**
@@ -111,23 +111,33 @@ class UsersAuthComponent extends Component
     {
         if (Configure::read('Users.auth')) {
             //initialize Auth
-            $this->_registry->getController()->loadComponent('Auth', Configure::read('Auth'));
+            $this->getController()->loadComponent('Auth', Configure::read('Auth'));
         }
 
-        $this->_registry->getController()->Auth->allow([
-            'register',
-            'validateEmail',
-            'resendTokenValidation',
-            'login',
-            'twitterLogin',
-            'socialEmail',
-            'resetPassword',
-            'requestResetPassword',
-            'changePassword',
-            'endpoint',
-            'authenticated',
-            'verify'
-        ]);
+        list($plugin, $controller) = pluginSplit(Configure::read('Users.controller'));
+        if ($this->getController()->request->getParam('plugin') === $plugin &&
+            $this->getController()->request->getParam('controller') === $controller
+        ) {
+            $this->getController()->Auth->allow([
+                // LoginTrait
+                'twitterLogin',
+                'login',
+                'socialEmail',
+                'verify',
+                // RegisterTrait
+                'register',
+                'validateEmail',
+                // PasswordManagementTrait used in RegisterTrait
+                'changePassword',
+                'resetPassword',
+                'requestResetPassword',
+                // UserValidationTrait used in PasswordManagementTrait
+                'resendTokenValidation',
+                // Social
+                'endpoint',
+                'authenticated',
+            ]);
+        }
     }
 
     /**
@@ -146,12 +156,12 @@ class UsersAuthComponent extends Component
 
         if (is_array($url)) {
             $requestUrl = Router::reverse($url);
-            $requestParams = Router::parse($requestUrl);
+            $requestParams = Router::parseRequest(new ServerRequest($requestUrl));
         } else {
             try {
                 //remove base from $url if exists
                 $normalizedUrl = Router::normalize($url);
-                $requestParams = Router::parse($normalizedUrl);
+                $requestParams = Router::parseRequest(new ServerRequest($normalizedUrl));
             } catch (MissingRouteException $ex) {
                 //if it's a url pointing to our own app
                 if (substr($normalizedUrl, 0, 1) === '/') {
@@ -168,15 +178,15 @@ class UsersAuthComponent extends Component
         }
 
         // check we are logged in
-        $user = $this->_registry->getController()->Auth->user();
+        $user = $this->getController()->Auth->user();
         if (empty($user)) {
             return false;
         }
 
-        $request = new Request($requestUrl);
-        $request->params = $requestParams;
+        $request = new ServerRequest($requestUrl);
+        $request = $request->addParams($requestParams);
 
-        $isAuthorized = $this->_registry->getController()->Auth->isAuthorized(null, $request);
+        $isAuthorized = $this->getController()->Auth->isAuthorized(null, $request);
 
         return $isAuthorized;
     }
@@ -206,7 +216,7 @@ class UsersAuthComponent extends Component
             return false;
         }
         $action = strtolower($requestParams['action']);
-        if (in_array($action, array_map('strtolower', $this->_registry->getController()->Auth->allowedActions))) {
+        if (in_array($action, array_map('strtolower', $this->getController()->Auth->allowedActions))) {
             return true;
         }