+++ src/Email/EmailSender.php
@@ -1,11 +1,11 @@
 <?php
 /**
- * Copyright 2010 - 2015, Cake Development Corporation (http://cakedc.com)
+ * Copyright 2010 - 2017, Cake Development Corporation (https://www.cakedc.com)
  *
  * Licensed under The MIT License
  * Redistributions of files must retain the above copyright notice.
  *
- * @copyright Copyright 2010 - 2015, Cake Development Corporation (http://cakedc.com)
+ * @copyright Copyright 2010 - 2017, Cake Development Corporation (https://www.cakedc.com)
  * @license MIT License (http://www.opensource.org/licenses/mit-license.php)
  */
 namespace CakeDC\Users\Email;
@@ -49,9 +49,12 @@ class EmailSender
      * instance
      * @return array email send result
      */
-    public function sendResetPasswordEmail(EntityInterface $user, Email $email = null, $template = 'CakeDC/Users.reset_password')
-    {
-        $this
+    public function sendResetPasswordEmail(
+        EntityInterface $user,
+        Email $email = null,
+        $template = 'CakeDC/Users.reset_password'
+    ) {
+        return $this
             ->getMailer(
                 'CakeDC/Users.Users',
                 $this->_getEmailInstance($email)
@@ -72,9 +75,10 @@ class EmailSender
         if (empty($email)) {
             $template = 'CakeDC/Users.social_account_validation';
         } else {
-            $template = $email->template()['template'];
+            $template = $email->getTemplate();
         }
-        $this
+
+        return $this
             ->getMailer(
                 'CakeDC/Users.Users',
                 $this->_getEmailInstance($email)
@@ -92,7 +96,7 @@ class EmailSender
     {
         if ($email === null) {
             $email = new Email('default');
-            $email->emailFormat('both');
+            $email->setEmailFormat('both');
         }
 
         return $email;