+++ src/Model/Behavior/BaseTokenBehavior.php
@@ -1,11 +1,11 @@
 <?php
 /**
- * Copyright 2010 - 2015, Cake Development Corporation (http://cakedc.com)
+ * Copyright 2010 - 2017, Cake Development Corporation (https://www.cakedc.com)
  *
  * Licensed under The MIT License
  * Redistributions of files must retain the above copyright notice.
  *
- * @copyright Copyright 2010 - 2015, Cake Development Corporation (http://cakedc.com)
+ * @copyright Copyright 2010 - 2017, Cake Development Corporation (https://www.cakedc.com)
  * @license MIT License (http://www.opensource.org/licenses/mit-license.php)
  */
 
@@ -13,12 +13,12 @@ namespace CakeDC\Users\Model\Behavior;
 
 use Cake\Datasource\EntityInterface;
 use Cake\I18n\Time;
-use Cake\ORM\Behavior as BaseBehavior;
+use Cake\ORM\Behavior;
 
 /**
  * Covers the user registration
  */
-class Behavior extends BaseBehavior
+class BaseTokenBehavior extends Behavior
 {
     /**
      * DRY for update active and token based on validateEmail flag