+++ src/View/Helper/UserHelper.php
@@ -1,11 +1,11 @@
 <?php
 /**
- * Copyright 2010 - 2015, Cake Development Corporation (http://cakedc.com)
+ * Copyright 2010 - 2017, Cake Development Corporation (https://www.cakedc.com)
  *
  * Licensed under The MIT License
  * Redistributions of files must retain the above copyright notice.
  *
- * @copyright Copyright 2010 - 2015, Cake Development Corporation (http://cakedc.com)
+ * @copyright Copyright 2010 - 2017, Cake Development Corporation (https://www.cakedc.com)
  * @license MIT License (http://www.opensource.org/licenses/mit-license.php)
  */
 
@@ -23,7 +23,6 @@ use Cake\View\Helper;
  */
 class UserHelper extends Helper
 {
-
     public $helpers = ['Html', 'Form', 'CakeDC/Users.AuthLink'];
 
     /**
@@ -55,7 +54,11 @@ class UserHelper extends Helper
             $providerTitle = __d('CakeDC/Users', '{0} {1}', Hash::get($options, 'label'), Inflector::camelize($name));
         }
 
-        $providerClass = __d('CakeDC/Users', 'btn btn-social btn-{0} ' . Hash::get($options, 'class') ?: '', strtolower($name));
+        $providerClass = __d(
+            'CakeDC/Users',
+            'btn btn-social btn-{0} ' . Hash::get($options, 'class') ?: '',
+            strtolower($name)
+        );
 
         return $this->Html->link($icon . $providerTitle, "/auth/$name", [
             'escape' => false, 'class' => $providerClass